Cleanup fixtures vs test utils
