+++ nbdime/tests/test_merge_notebooks.py
@@ -11,8 +11,7 @@ from six import string_types
 import nbformat
 
 from nbdime.diff_format import op_patch, op_addrange, op_removerange, op_replace
-from .conftest import have_git
-from .fixtures import sources_to_notebook, matching_nb_triplets, outputs_to_notebook
+from .fixtures import sources_to_notebook, outputs_to_notebook, have_git
 from nbdime.nbmergeapp import _build_arg_parser
 from nbdime import merge_notebooks, apply_decisions
 from nbdime.diffing.notebooks import diff_notebooks
