+++ nbdime/tests/test_web.py
@@ -12,15 +12,12 @@ from tornado.escape import json_encode, json_decode
 import nbformat
 import json
 
-from .fixtures import filespath
-
 import nbdime.webapp.nbdiffweb
 import nbdime.webapp.nbmergeweb
 
 WEB_TEST_TIMEOUT = 15
 
 
-files = filespath()
 diff_a = 'src-and-output--1.ipynb'
 diff_b = 'src-and-output--2.ipynb'
 
@@ -28,23 +25,23 @@ merge_a = 'multilevel-test-base.ipynb'
 merge_b = 'multilevel-test-local.ipynb'
 merge_c = 'multilevel-test-remote.ipynb'
 
+
 @pytest.mark.timeout(timeout=WEB_TEST_TIMEOUT)
-def test_diff_web():
+def test_diff_web(filespath):
     port = 62023
-    a = os.path.join(files, diff_a)
-    b = os.path.join(files, diff_b)
+    a = os.path.join(filespath, diff_a)
+    b = os.path.join(filespath, diff_b)
     loop = ioloop.IOLoop.current()
     loop.call_later(0, loop.stop)
     nbdime.webapp.nbdiffweb.main(['--port=%i' % port, '--browser=disabled', a, b])
 
 
 @pytest.mark.timeout(timeout=WEB_TEST_TIMEOUT)
-def test_merge_web():
+def test_merge_web(filespath):
     port = 62024
-    files = filespath()
-    a = os.path.join(files, merge_a)
-    b = os.path.join(files, merge_b)
-    c = os.path.join(files, merge_c)
+    a = os.path.join(filespath, merge_a)
+    b = os.path.join(filespath, merge_b)
+    c = os.path.join(filespath, merge_c)
     loop = ioloop.IOLoop.current()
     loop.call_later(0, loop.stop)
     nbdime.webapp.nbmergeweb.main(['--port=%i' % port, '--browser=disabled', a, b, c])
@@ -62,7 +59,7 @@ def test_fetch_diff(http_client, base_url, nbdime_base_url):
 
 @pytest.mark.timeout(timeout=WEB_TEST_TIMEOUT)
 @pytest.mark.gen_test
-def test_api_diff(http_client, base_url, nbdime_base_url, diff_validator):
+def test_api_diff(http_client, base_url, nbdime_base_url, diff_validator, filespath):
     post_data = dict(base=diff_a, remote=diff_b)
     body = json_encode(post_data)
 
@@ -72,7 +69,7 @@ def test_api_diff(http_client, base_url, nbdime_base_url, diff_validator):
     # Check that response is sane:
     data = json_decode(response.body)
     # Check that base is as expected:
-    expected_base = nbformat.read(os.path.join(files, diff_a), as_version=4)
+    expected_base = nbformat.read(os.path.join(filespath, diff_a), as_version=4)
     assert json.dumps(data['base'], sort_keys=True) == json.dumps(expected_base, sort_keys=True)
     # Check that diff follows schema:
     diff_validator.validate(data['diff'])
@@ -90,7 +87,7 @@ def test_fetch_merge(http_client, base_url, nbdime_base_url):
 
 @pytest.mark.timeout(timeout=WEB_TEST_TIMEOUT)
 @pytest.mark.gen_test
-def test_api_merge(http_client, base_url, nbdime_base_url, merge_validator):
+def test_api_merge(http_client, base_url, nbdime_base_url, merge_validator, filespath):
     post_data = dict(base=merge_a, local=merge_b, remote=merge_c)
     body = json_encode(post_data)
 
@@ -100,7 +97,7 @@ def test_api_merge(http_client, base_url, nbdime_base_url, merge_validator):
     # Check that response is sane:
     data = json_decode(response.body)
     # Check that base is as expected:
-    expected_base = nbformat.read(os.path.join(files, merge_a), as_version=4)
+    expected_base = nbformat.read(os.path.join(filespath, merge_a), as_version=4)
     assert json.dumps(data['base'], sort_keys=True) == json.dumps(expected_base, sort_keys=True)
     # Check that decisions follows schema:
     merge_validator.validate(data['merge_decisions'])
